+++ src/main/java/com/github/wintersteve25/tau/components/Renderable.java
@@ -1,8 +1,8 @@
 package com.github.wintersteve25.tau.components;
 
 import com.github.wintersteve25.tau.theme.Theme;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.client.gui.IRenderable;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.components.Widget;
 import com.github.wintersteve25.tau.components.base.DynamicUIComponent;
 import com.github.wintersteve25.tau.components.base.PrimitiveUIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
@@ -12,14 +12,14 @@ import java.util.List;
 
 public final class Renderable implements PrimitiveUIComponent {
     
-    private final IRenderable renderable;
+    private final Widget renderable;
 
-    public Renderable(IRenderable renderable) {
+    public Renderable(Widget renderable) {
         this.renderable = renderable;
     }
 
     @Override
-    public Vector2i build(Layout layout, Theme theme, List<IRenderable> renderables, List<IRenderable> tooltips, List<DynamicUIComponent> dynamicUIComponents, List<IGuiEventListener> eventListeners) {
+    public Vector2i build(Layout layout, Theme theme, List<Widget> renderables, List<Widget> tooltips, List<DynamicUIComponent> dynamicUIComponents, List<GuiEventListener> eventListeners) {
         renderables.add(renderable);
         return layout.getSize();
     }