+++ src/main/java/com/github/wintersteve25/tau/tests/TestStack.java
@@ -4,10 +4,9 @@ import com.github.wintersteve25.tau.components.*;
 import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
 import com.github.wintersteve25.tau.theme.Theme;
-import com.github.wintersteve25.tau.utils.Color;
 import com.github.wintersteve25.tau.utils.Size;
 import com.github.wintersteve25.tau.utils.Vector2i;
-import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.network.chat.TextComponent;
 
 public class TestStack implements UIComponent {
     @Override
@@ -17,8 +16,8 @@ public class TestStack implements UIComponent {
                 new Center(new Sized(
                         Size.staticSize(new Vector2i(100, 20)),
                         new TextField.Builder()
-                                .withMessage(new StringTextComponent("Hello"))
-                                .withHintText(new StringTextComponent("Hello!")))
+                                .withMessage(new TextComponent("Hello"))
+                                .withHintText(new TextComponent("Hello!")))
                 ));
     }
 }