+++ src/main/java/com/github/wintersteve25/tau/theme/Theme.java
@@ -2,20 +2,20 @@ package com.github.wintersteve25.tau.theme;
 
 import com.github.wintersteve25.tau.utils.Color;
 import com.github.wintersteve25.tau.utils.InteractableState;
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.gui.Font;
+import net.minecraft.network.chat.FormattedText;
+
+import java.util.List;
 
 public interface Theme {
-    void drawButton(MatrixStack matrixStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY, InteractableState state);
+    void drawButton(PoseStack postStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY, InteractableState state);
     
-    void drawContainer(MatrixStack matrixStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY);
+    void drawContainer(PoseStack poseStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY);
     
-    void drawScrollbar(MatrixStack matrixStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY);
+    void drawScrollbar(PoseStack poseStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY);
+   
+    void drawTooltip(PoseStack poseStack, int mouseX, int mouseY, int screenWidth, int screenHeight, Font fontRenderer, List<FormattedText> tooltips);
     
     Color getTextColor();
-    
-    Color getTooltipColor();
-    
-    Color getTooltipBorderStartColor();
-    
-    Color getTooltipBorderEndColor();
 }