+++ src/main/java/com/github/wintersteve25/tau/utils/transformations/Transformation.java
@@ -1,7 +1,7 @@
 package com.github.wintersteve25.tau.utils.transformations;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
 
 public interface Transformation {
-    void transform(MatrixStack matrixStack);
+    void transform(PoseStack PoseStack);
 }